+++ cpp/dolfinx/fem/CMakeLists.txt
@@ -30,7 +30,6 @@ target_sources(
   ${CMAKE_CURRENT_SOURCE_DIR}/DofMap.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/ElementDofLayout.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/FiniteElement.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/FunctionSpace.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/dofmapbuilder.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/petsc.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/sparsitybuild.cpp
